requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-generativeai==0.8.5
pydantic==2.7.1
python-dotenv==1.0.0
chromadb==1.0.15
sentence-transformers==5.0.0
aiofiles==23.2.1
python-multipart==0.0.6
websockets==12.0
jinja2==3.1.2
markdown==3.5.1
requests==2.31.0
numpy==1.26.4
setuptools>=65.0.0
typing-extensions>=4.5.0
pydantic-settings==2.3.4 